rename categories section appropriately
