--- lib/screens/home/component/body.dart
@@ -1,22 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:mobileapp/screens/home/component/top_section.dart';
-
-import 'content.dart';
-
-class HomeBody extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 10.0),
-      child: Column(
-        children: [
-          TopSection(),
-          SizedBox(
-            height: 15.0,
-          ),
-          BodyContent(),
-        ],
-      ),
-    );
-  }
-}
