+++ lib/screens/home/home.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:mobileapp/components/bottom_navigation.dart';
 import 'package:mobileapp/components/custom_appbar.dart';
-import 'package:mobileapp/screens/home/component/body.dart';
+import 'package:mobileapp/screens/home/component/categories_section.dart';
+import 'package:mobileapp/screens/home/component/content.dart';
 import '../../utils/enum.dart';
 
 class HomeScreen extends StatelessWidget {
@@ -14,7 +15,18 @@ class HomeScreen extends StatelessWidget {
       ),
       body: Container(
         height: MediaQuery.of(context).size.height,
-        child: HomeBody(),
+        child: Padding(
+          padding: const EdgeInsets.symmetric(horizontal: 10.0),
+          child: Column(
+            children: [
+              CategoriesSection(),
+              SizedBox(
+                height: 15.0,
+              ),
+              BodyContent(),
+            ],
+          ),
+        ),
       ),
       bottomNavigationBar: BottomNavigation(
         selectedMenu: MenuState.home,
